Fix a mistake of `feeGrowthOutside` initialization
The piecewise function in the chapter "不同区间之间的分配" mentions the rules of `feeGrowthOutside` initialization. I discovered that the formula given in the chapter is the exact opposite of the formula shown in the whitepaper of Uniswap V3(i.e., formula 6.21).

The assumption of `feeGrowthOutside` in the whitepaper is described as follows:
"When 𝑓𝑜 is initialized for a tick 𝑖, the value—by convention—is
chosen as if all of the fees earned to date had occurred below that
tick".
This means `fee_global` ought to be at the left side of the tick which is going to be initialized. Therefore, when `i_current` is greater than or equal to `i`, the 'outside' under this case is the left side of the tick(the value of `feeGrowthOutside` should be `fee_global` instead of `0`). I swapped the two values in the piecewise function to correct the mistake currently shown.